Phase 3.4: Split play-view-drawer.tsx into orchestrator + 4 siblings
play-view-drawer.tsx had grown to ~1500 lines, hosting an ~800-line
orchestrator plus three already-memoised sub-components (PlayDrawerContent,
PlayViewActionBar, PlayViewTickBar) and an inline wall-view banner
hard-wired into the parent's `aboveFold` useMemo dep array.

Extract each sub-component to its own sibling file:
- play-drawer-content.tsx     — PlayDrawerContent
- play-view-action-bar.tsx    — PlayViewActionBar + props
- play-view-tick-bar.tsx      — PlayViewTickBar + props
- wall-view-banner.tsx        — WallViewBanner (was inline `Box` in the
                                aboveFold memo; now its own component
                                that reads `t` directly, so the memo's
                                wall-view branch is a 1-liner)

Update test imports: play-view-action-bar.test.tsx and
play-view-tick-bar-overlay.test.tsx now point to the new sibling files
rather than the orchestrator's re-exports. No compatibility shim — the
extraction is the point.

Trim the orchestrator's import list: MuiBadge, Tooltip, TextField,
InputAdornment, the icon-bar icons, QuickTickBar, useBuildClimbDetailSections,
useLongPress, hasPriorHistoryForClimb, getGradeTintColor, useIsDarkMode,
TickIcon, TickButtonWithLabel, PersonFallingIcon, ChatBubbleOutlineOutlined,
KeyboardArrow*Outlined, Box, Favorite, Lightbulb, and the Climb type all
moved out with their respective sub-components.

Net: orchestrator drops from 1519 → 1043 lines (−476), four new files
totaling ~530 lines. Same code overall, just split where the seams already
were. The action bar tests still get the same component, the tick-bar
tests still get the same component, and party + solo flows in the
drawer are byte-identical at the React tree level.
