+++ packages/web/app/components/play-view/play-view-drawer.tsx
@@ -3,31 +3,12 @@
 import React, { useCallback, useEffect, useLayoutEffect, useRef, useState, useMemo, useDeferredValue } from 'react';
 import { useTranslation } from 'react-i18next';
 import { track } from '@/app/lib/analytics';
-import MuiBadge from '@mui/material/Badge';
 import Box from '@mui/material/Box';
 import IconButton from '@mui/material/IconButton';
-import { TickBadgeAvatar } from '@/app/components/session/tick-badge-avatar';
-import { MiniSessionBar } from './mini-session-bar';
-import Tooltip from '@mui/material/Tooltip';
-import TextField from '@mui/material/TextField';
-import InputAdornment from '@mui/material/InputAdornment';
-import SyncOutlined from '@mui/icons-material/SyncOutlined';
-import FavoriteBorderOutlined from '@mui/icons-material/FavoriteBorderOutlined';
-import Favorite from '@mui/icons-material/Favorite';
-import SkipPreviousOutlined from '@mui/icons-material/SkipPreviousOutlined';
-import SkipNextOutlined from '@mui/icons-material/SkipNextOutlined';
-import LightbulbOutlined from '@mui/icons-material/LightbulbOutlined';
-import Lightbulb from '@mui/icons-material/Lightbulb';
-import MoreHorizOutlined from '@mui/icons-material/MoreHorizOutlined';
 import CloseOutlined from '@mui/icons-material/CloseOutlined';
-import KeyboardArrowUpOutlined from '@mui/icons-material/KeyboardArrowUpOutlined';
-import KeyboardArrowDownOutlined from '@mui/icons-material/KeyboardArrowDownOutlined';
-import FormatListBulletedOutlined from '@mui/icons-material/FormatListBulletedOutlined';
 import CheckOutlined from '@mui/icons-material/CheckOutlined';
-import ChatBubbleOutlineOutlined from '@mui/icons-material/ChatBubbleOutlineOutlined';
-import { TickIcon, TickButtonWithLabel } from '../logbook/tick-icon';
+import { TickBadgeAvatar } from '@/app/components/session/tick-badge-avatar';
 import type { ClimbQueueItem } from '../queue-control/types';
-import { PersonFallingIcon } from '@/app/components/icons/person-falling-icon';
 import { usePathname } from 'next/navigation';
 import { useQueueActions, useCurrentClimb, useQueueList, useSessionData } from '../graphql-queue';
 import { ClimbActions } from '../climb-actions';
@@ -36,7 +17,6 @@ import HeartAnimationOverlay from '../climb-card/heart-animation-overlay';
 import PlaylistSelectionContent from '../climb-actions/playlist-selection-content';
 import DrawerClimbHeader from '../climb-card/drawer-climb-header';
 import { LightControlDrawer } from '../board-page/light-control-drawer';
-import { useLongPress } from '@/app/lib/hooks/use-long-press';
 import { useBoardProvider } from '../board-provider/board-provider-context';
 import SwipeBoardCarousel from '../board-renderer/swipe-board-carousel';
 import { useWakeLock } from '../board-bluetooth-control/use-wake-lock';
@@ -46,498 +26,33 @@ import { themeTokens } from '@/app/theme/theme-config';
 import SwipeableDrawer from '../swipeable-drawer/swipeable-drawer';
 import AngleSelector from '../board-page/angle-selector';
 import ClimbDetailHeader from '@/app/components/climb-detail/climb-detail-header';
-import { QuickTickBar, type QuickTickBarHandle } from '../logbook/quick-tick-bar';
-import { hasPriorHistoryForClimb } from '@/app/hooks/use-tick-save';
 import type { ActiveDrawer } from '../queue-control/queue-control-bar';
 import { PLAY_DRAWER_EVENT } from '../queue-control/play-drawer-event';
 import {
   TOUR_CLOSE_PLAY_QUEUE_EVENT,
   TOUR_OPEN_PLAY_QUEUE_EVENT,
 } from '@/app/components/onboarding/onboarding-tour-events';
-import type { BoardDetails, Angle, Climb } from '@/app/lib/types';
+import type { BoardDetails, Angle } from '@/app/lib/types';
 import styles from './play-view-drawer.module.css';
 import drawerCss from '../swipeable-drawer/swipeable-drawer.module.css';
 import { useDrawerDragResize } from '@/app/hooks/use-drawer-drag-resize';
 import ClimbDetailShellClient from '@/app/components/climb-detail/climb-detail-shell.client';
-import { useBuildClimbDetailSections } from '@/app/components/climb-detail/build-climb-detail-sections';
 import { renderBoard } from '@/app/lib/board-render-worker/worker-manager';
 import { useNestedDrawerSwipe } from '@/app/lib/hooks/use-nested-drawer-swipe';
 import { usePullToClose, findScrollContainer } from '@/app/lib/hooks/pull-to-close';
 import { useSnackbar } from '../providers/snackbar-provider';
-import { getGradeTintColor } from '@/app/lib/grade-colors';
-import { useIsDarkMode } from '@/app/hooks/use-is-dark-mode';
 import { getPreference, setPreference } from '@/app/lib/user-preferences-db';
 import QueueDrawer from './queue-drawer';
+import PlayDrawerContent from './play-drawer-content';
+import { PlayViewActionBar } from './play-view-action-bar';
+import { PlayViewTickBar } from './play-view-tick-bar';
+import { MiniSessionBar } from './mini-session-bar';
 import { useDrawerUrlSync } from './use-drawer-url-sync';
 
 /** Window with optional requestIdleCallback (not available in all browsers). */
 type WindowWithIdleCallback = Window & {
   requestIdleCallback?: ((cb: () => void, opts?: { timeout: number }) => number) | undefined;
 };
-
-type PlayDrawerContentProps = {
-  climb: Climb;
-  boardType: string;
-  angle: number;
-  layoutId: number;
-  viewerBoardDetails: BoardDetails;
-  aboveFold: React.ReactNode;
-  sectionsEnabled: boolean;
-  paperRef: React.RefObject<HTMLDivElement | null>;
-};
-
-const PlayDrawerContent = React.memo<PlayDrawerContentProps>(
-  ({ climb, boardType, angle, layoutId, viewerBoardDetails, aboveFold, sectionsEnabled, paperRef }) => {
-    const sections = useBuildClimbDetailSections({
-      climb,
-      climbUuid: climb.uuid,
-      boardType,
-      angle,
-      layoutId,
-      viewerBoardDetails,
-      currentClimbDifficulty: climb.difficulty ?? undefined,
-      boardName: boardType,
-      enabled: sectionsEnabled,
-    });
-
-    // When the active climb changes (e.g. tapping a card in the Similar
-    // climbs slider activates a new climb), reset the drawer's scroll
-    // position to the top so the user sees the new climb's header /
-    // board first instead of landing mid-scroll on whatever section
-    // they were on in the previous climb. The data-scroll-container
-    // attribute is set by ClimbDetailShellClient on its mobileScrollLayout
-    // wrapper (climb-detail-shell.client.tsx:38). Scope to paperRef so we
-    // don't accidentally grab a foreign element that happens to carry the
-    // same attribute elsewhere on the page.
-    useEffect(() => {
-      const scrollContainer = paperRef.current?.querySelector<HTMLElement>('[data-scroll-container]');
-      if (scrollContainer) {
-        scrollContainer.scrollTo({ top: 0, behavior: 'smooth' });
-      }
-    }, [climb.uuid, paperRef]);
-
-    return <ClimbDetailShellClient mode="play" sections={sections} aboveFold={aboveFold} />;
-  },
-);
-PlayDrawerContent.displayName = 'PlayDrawerContent';
-
-type PlayViewActionBarProps = {
-  canSwipePrevious: boolean;
-  canSwipeNext: boolean;
-  isMirrored: boolean;
-  supportsMirroring: boolean;
-  isFavorited: boolean;
-  remainingQueueCount: number;
-  onPrevClick: () => void;
-  onNextClick: () => void;
-  onMirror: () => void;
-  onToggleFavorite: () => void;
-  onOpenActions: () => void;
-  onOpenQueue: () => void;
-  /** Whether the lightbulb should render as filled/lit. In party this is
-   *  `isDriver` (you hold the wall). In solo it's `isBluetoothConnected`
-   *  (a paired board IS the wall — when nothing's paired, the lightbulb is
-   *  outlined to signal "tap to connect"). The driver concept doesn't carry
-   *  any BLE meaning in solo, so the visual would always be lit if we used
-   *  isDriver there — masking the actual connection state. */
-  lightbulbActive: boolean;
-  /** Pulse the lightbulb while a take-control press is in flight. Set
-   *  between the press and the matching `WallConfirmedClimb` event (or the
-   *  2-second timeout fallback). Independent from `lightbulbActive`. */
-  lightbulbPending?: boolean;
-  /** Single-iteration pulse + tooltip on first drawer open. The parent
-   *  reads `swipeHint:lightbulbSeen` from IndexedDB and toggles this true
-   *  once, then writes the flag so subsequent opens skip the coachmark. */
-  lightbulbCoachmark?: boolean;
-  /** Localized tooltip copy for the coachmark. Threaded through so the
-   *  parent owns the i18n call and this component stays presentational. */
-  lightbulbCoachmarkText?: string;
-  /** Fired when the coachmark animation runs once — the parent persists
-   *  the seen flag and clears `lightbulbCoachmark`. */
-  onLightbulbCoachmarkSeen?: () => void;
-  /** Name of the currently displayed climb. Used in the lightbulb's aria
-   *  label so screen-reader users hear what they're sending. */
-  displayedClimbName: string | null;
-  /** Pivot's lightbulb gesture: in solo it sends the climb to the wall via
-   *  the existing BLE auto-sender path; in party it claims driver and
-   *  broadcasts the climb. */
-  onLightbulb: () => void;
-  /** Long-press the lightbulb to open the light-control drawer (disco, party
-   *  glyphs, palette, and the manual BLE disconnect). Optional — when the
-   *  parent doesn't supply it, long-press is a no-op. */
-  onLightbulbLongPress?: () => void;
-  angleSelector?: React.ReactNode;
-};
-
-export const PlayViewActionBar = React.memo(function PlayViewActionBar({
-  canSwipePrevious,
-  canSwipeNext,
-  isMirrored,
-  supportsMirroring,
-  isFavorited,
-  remainingQueueCount,
-  onPrevClick,
-  onNextClick,
-  onMirror,
-  onToggleFavorite,
-  onOpenActions,
-  onOpenQueue,
-  lightbulbActive,
-  lightbulbPending = false,
-  lightbulbCoachmark = false,
-  lightbulbCoachmarkText,
-  onLightbulbCoachmarkSeen,
-  displayedClimbName,
-  onLightbulb,
-  onLightbulbLongPress,
-  angleSelector,
-}: PlayViewActionBarProps) {
-  const { t } = useTranslation('session');
-  // Lightbulb aria label — active (filled) vs inactive (outlined) framing
-  // makes the action's destructive-vs-additive nature explicit for screen
-  // readers. "Driving" copy reads fine in solo too (you're driving the board
-  // when BLE is connected).
-  const lightbulbLabel = displayedClimbName
-    ? lightbulbActive
-      ? t('playView.actionBar.lightbulb.drivingNamed', { name: displayedClimbName })
-      : t('playView.actionBar.lightbulb.takeNamed', { name: displayedClimbName })
-    : lightbulbActive
-      ? t('playView.actionBar.lightbulb.driving')
-      : t('playView.actionBar.lightbulb.take');
-  // Long-press the lightbulb to reach the light-control drawer (and the
-  // manual BLE disconnect that lives inside it). Tap stays the take-control
-  // gesture; consumeLongPress() in the click handler swallows the synthesized
-  // click that follows a long-press so we don't fire both.
-  const { ref: lightbulbLongPressRef, consumeLongPress } = useLongPress<HTMLButtonElement>(onLightbulbLongPress);
-  const handleLightbulbTap = useCallback(() => {
-    if (consumeLongPress()) return;
-    onLightbulb();
-  }, [consumeLongPress, onLightbulb]);
-  return (
-    <div className={styles.actionBar}>
-      <IconButton disabled={!canSwipePrevious} onClick={onPrevClick}>
-        <SkipPreviousOutlined />
-      </IconButton>
-      {supportsMirroring && (
-        <IconButton
-          color={isMirrored ? 'primary' : 'default'}
-          onClick={onMirror}
-          sx={
-            isMirrored
-              ? {
-                  backgroundColor: themeTokens.colors.purple,
-                  borderColor: themeTokens.colors.purple,
-                  color: 'common.white',
-                  '&:hover': { backgroundColor: themeTokens.colors.purple },
-                }
-              : undefined
-          }
-        >
-          <SyncOutlined />
-        </IconButton>
-      )}
-      <IconButton onClick={onToggleFavorite}>
-        {isFavorited ? <Favorite sx={{ color: themeTokens.colors.error }} /> : <FavoriteBorderOutlined />}
-      </IconButton>
-      {/* Lightbulb: the queue-control-bar pivot's primary "send/take" gesture.
-          Filled+amber-glowing when the lightbulb is active (driver in party,
-          BLE-paired in solo); outlined when inactive (non-driver in party,
-          unpaired in solo). The warm-amber styling matches the ShareBoardButton
-          that this drawer replaced — it reads as "this bulb is lit" rather
-          than the dusty-rose primary, which the user kept misreading as an
-          error state. Long-press opens the light-control drawer (disco /
-          glyphs / palette / manual disconnect).
-
-          Pending / coachmark variants share a single CSS pulse class so the
-          two states never compose into a janky doubled animation. */}
-      <Tooltip
-        // Coachmark tooltip only — once `lightbulbCoachmark` flips false the
-        // tooltip is gone (default-closed Tooltip with no children-trigger).
-        open={lightbulbCoachmark && !!lightbulbCoachmarkText}
-        title={lightbulbCoachmarkText ?? ''}
-        placement="top"
-        arrow
-        // Tooltip-only — the underlying button still owns its aria-label.
-        disableInteractive
-      >
-        <IconButton
-          ref={lightbulbLongPressRef}
-          onClick={handleLightbulbTap}
-          aria-label={lightbulbLabel}
-          className={
-            lightbulbPending ? styles.lightbulbPending : lightbulbCoachmark ? styles.lightbulbCoachmark : undefined
-          }
-          onAnimationEnd={lightbulbCoachmark ? onLightbulbCoachmarkSeen : undefined}
-        >
-          {lightbulbActive ? (
-            <Lightbulb className={styles.lightbulbConnectedGlow} sx={{ color: themeTokens.colors.warning }} />
-          ) : (
-            <LightbulbOutlined />
-          )}
-        </IconButton>
-      </Tooltip>
-      {angleSelector}
-      <IconButton onClick={onOpenActions} aria-label={t('playView.actionBar.climbActionsAria')}>
-        <MoreHorizOutlined />
-      </IconButton>
-      <MuiBadge
-        badgeContent={remainingQueueCount}
-        max={99}
-        sx={{
-          '& .MuiBadge-badge': {
-            backgroundColor: themeTokens.colors.primary,
-            color: 'common.white',
-          },
-        }}
-      >
-        <IconButton onClick={onOpenQueue} aria-label={t('playView.actionBar.openQueueAria')}>
-          <FormatListBulletedOutlined />
-        </IconButton>
-      </MuiBadge>
-      <IconButton disabled={!canSwipeNext} onClick={onNextClick}>
-        <SkipNextOutlined />
-      </IconButton>
-    </div>
-  );
-});
-PlayViewActionBar.displayName = 'PlayViewActionBar';
-
-/**
- * Extracted tick bar component that owns its own `tickComment` state.
- * This prevents comment keystrokes from invalidating the parent `aboveFold` useMemo,
- * which would otherwise re-render the entire board carousel on every keystroke.
- */
-type PlayViewTickBarProps = {
-  isTickBarActive: boolean;
-  currentClimb: Climb;
-  angle: Angle;
-  boardDetails: BoardDetails;
-  onClose: () => void;
-  onError: () => void;
-};
-
-export const PlayViewTickBar = React.memo<PlayViewTickBarProps>(function PlayViewTickBar({
-  isTickBarActive,
-  currentClimb,
-  angle,
-  boardDetails,
-  onClose,
-  onError,
-}) {
-  const { t } = useTranslation('session');
-  const { logbook } = useBoardProvider();
-  const [tickComment, setTickComment] = useState('');
-  const [commentFocused, setCommentFocused] = useState(false);
-  const [isFlash, setIsFlash] = useState(() => !hasPriorHistoryForClimb(currentClimb, logbook));
-  const [tickBarExpanded, setTickBarExpanded] = useState(false);
-  const quickTickBarRef = useRef<QuickTickBarHandle>(null);
-  const isDark = useIsDarkMode();
-  // Opaque surface base + grade tint overlay (matches queue control bar pattern).
-  const gradeTintColor = useMemo(
-    () => getGradeTintColor(currentClimb.difficulty, 'default', isDark),
-    [currentClimb.difficulty, isDark],
-  );
-
-  // Restore persisted tick bar expanded state when tick bar opens
-  useEffect(() => {
-    if (isTickBarActive) {
-      void getPreference<boolean>('tickBarExpanded').then((persisted) => {
-        if (persisted === true) setTickBarExpanded(true);
-      });
-    }
-  }, [isTickBarActive]);
-
-  // Persist expanded state on user-initiated toggle (not on automatic resets)
-  const handleTickBarExpandedChange = useCallback((expanded: boolean) => {
-    setTickBarExpanded(expanded);
-    void setPreference('tickBarExpanded', expanded);
-  }, []);
-
-  const handleCommentFocus = useCallback(() => setCommentFocused(true), []);
-  const handleCommentBlur = useCallback(() => setCommentFocused(false), []);
-
-  // Reset comment when the tick bar closes
-  const handleClose = useCallback(() => {
-    setTickComment('');
-    setCommentFocused(false);
-    setIsFlash(false);
-    setTickBarExpanded(false);
-    onClose();
-  }, [onClose]);
-
-  // Reset comment and recompute flash state when the climb changes.
-  useEffect(() => {
-    setTickComment('');
-    setCommentFocused(false);
-    setIsFlash(!hasPriorHistoryForClimb(currentClimb, logbook));
-    setTickBarExpanded(false);
-  }, [currentClimb, logbook]);
-
-  return (
-    <div className={`${styles.tickBarContainer} ${isTickBarActive ? styles.tickBarContainerActive : ''}`}>
-      <div
-        className={styles.tickBarInner}
-        style={{
-          backgroundColor: isDark ? 'var(--semantic-surfaceElevated)' : 'var(--semantic-surface)',
-          // Grade tint as a solid overlay via linear-gradient (single-color gradient)
-          ...(gradeTintColor ? { backgroundImage: `linear-gradient(${gradeTintColor}, ${gradeTintColor})` } : {}),
-        }}
-      >
-        {isTickBarActive && (
-          <>
-            {/* Toolbar: expand left, close right — same pattern as queue-control-bar */}
-            <div className={styles.tickBarToolbar}>
-              <div
-                className={styles.tickBarExpandButton}
-                onClick={() => handleTickBarExpandedChange(!tickBarExpanded)}
-                role="button"
-                tabIndex={0}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter' || e.key === ' ') handleTickBarExpandedChange(!tickBarExpanded);
-                }}
-                aria-label={tickBarExpanded ? t('playView.tickBar.collapseAria') : t('playView.tickBar.expandAria')}
-              >
-                {tickBarExpanded ? (
-                  <KeyboardArrowDownOutlined sx={{ fontSize: 16, opacity: 0.7 }} />
-                ) : (
-                  <KeyboardArrowUpOutlined sx={{ fontSize: 16, opacity: 0.7 }} />
-                )}
-                <span className={styles.tickBarExpandLabel}>
-                  {tickBarExpanded ? t('queueBar.tickBar.collapse') : t('queueBar.tickBar.expand')}
-                </span>
-              </div>
-              <div className={styles.tickBarCloseButton}>
-                <IconButton
-                  size="small"
-                  onClick={handleClose}
-                  aria-label={t('playView.tickBar.closeAria')}
-                  sx={{
-                    color: 'text.primary',
-                    backgroundColor: 'action.selected',
-                    '&:hover': { backgroundColor: 'action.focus' },
-                    padding: '2px',
-                  }}
-                >
-                  <CloseOutlined sx={{ fontSize: 16 }} />
-                </IconButton>
-              </div>
-            </div>
-            <QuickTickBar
-              ref={quickTickBarRef}
-              currentClimb={currentClimb}
-              angle={angle}
-              boardDetails={boardDetails}
-              onSave={handleClose}
-              onError={onError}
-              onDraftRestored={(draftComment) => setTickComment(draftComment)}
-              onIsFlashChange={setIsFlash}
-              comment={tickComment}
-              expanded={tickBarExpanded}
-              commentSlot={
-                <div className={`${styles.tickBarComment} ${commentFocused ? styles.tickBarCommentExpanded : ''}`}>
-                  <TextField
-                    fullWidth
-                    size="small"
-                    variant="outlined"
-                    placeholder={t('common:comment.shortPlaceholder')}
-                    multiline
-                    minRows={1}
-                    maxRows={commentFocused ? 4 : 1}
-                    value={tickComment}
-                    onChange={(e) => setTickComment(e.target.value)}
-                    onFocus={handleCommentFocus}
-                    onBlur={handleCommentBlur}
-                    slotProps={{
-                      htmlInput: { maxLength: 2000, 'aria-label': 'Tick comment' },
-                      input: {
-                        startAdornment: (
-                          <InputAdornment position="start">
-                            <ChatBubbleOutlineOutlined sx={{ fontSize: 16, opacity: 0.5 }} />
-                          </InputAdornment>
-                        ),
-                      },
-                    }}
-                    sx={{
-                      '& .MuiOutlinedInput-root': {
-                        borderRadius: '8px',
-                        backgroundColor: 'var(--input-bg)',
-                        '& .MuiOutlinedInput-notchedOutline': {
-                          borderColor: 'var(--neutral-200)',
-                        },
-                      },
-                    }}
-                  />
-                </div>
-              }
-              expandedCommentSlot={
-                <TextField
-                  fullWidth
-                  size="small"
-                  variant="outlined"
-                  placeholder={t('playView.tickBar.commentPlaceholder')}
-                  multiline
-                  minRows={2}
-                  maxRows={4}
-                  value={tickComment}
-                  onChange={(e) => setTickComment(e.target.value)}
-                  onFocus={handleCommentFocus}
-                  onBlur={handleCommentBlur}
-                  slotProps={{
-                    htmlInput: { maxLength: 2000, 'aria-label': 'Tick comment' },
-                  }}
-                  sx={{
-                    '& .MuiOutlinedInput-root': {
-                      borderRadius: '8px',
-                      backgroundColor: 'var(--input-bg)',
-                      '& .MuiOutlinedInput-notchedOutline': {
-                        borderColor: 'var(--neutral-200)',
-                      },
-                    },
-                  }}
-                />
-              }
-            />
-            {/* Action buttons — attempt + tick (order matches queue control bar) */}
-
-            <div className={styles.tickBarButtons}>
-              <TickButtonWithLabel label={t('playView.tickBar.attemptLabel')}>
-                <IconButton
-                  onClick={(e) => quickTickBarRef.current?.saveAttempt(e.currentTarget)}
-                  sx={{
-                    backgroundColor: themeTokens.colors.errorMuted,
-                    color: themeTokens.colors.error,
-                    '&:hover': { backgroundColor: themeTokens.colors.errorMutedHover },
-                  }}
-                  aria-label={t('playView.tickBar.logAttemptAria')}
-                >
-                  <PersonFallingIcon />
-                </IconButton>
-              </TickButtonWithLabel>
-              <TickButtonWithLabel label={isFlash ? 'flash' : 'tick'}>
-                <IconButton
-                  id="button-tick"
-                  onClick={(e) => quickTickBarRef.current?.save(e.currentTarget)}
-                  sx={{
-                    backgroundColor: isFlash ? themeTokens.colors.amber : themeTokens.colors.success,
-                    color: isFlash ? themeTokens.neutral[900] : 'common.white',
-                    transition: 'background-color 150ms ease, color 150ms ease',
-                    '&:hover': {
-                      backgroundColor: isFlash ? themeTokens.colors.amber : themeTokens.colors.successHover,
-                    },
-                  }}
-                  aria-label={t('playView.tickBar.saveTickAria')}
-                >
-                  <TickIcon isFlash={!!isFlash} />
-                </IconButton>
-              </TickButtonWithLabel>
-            </div>
-          </>
-        )}
-      </div>
-    </div>
-  );
-});
-PlayViewTickBar.displayName = 'PlayViewTickBar';
 
 type PlayViewDrawerProps = {
   activeDrawer: ActiveDrawer;
